Fix broken references and rename `to_requirements_string` to `to_requirement_string`
