+++ tests/usethis/_types/test_deps.py
@@ -5,12 +5,12 @@ class TestDependency:
     class TestToRequirementsString:
         def test_no_extras(self):
             dep = Dependency(name="requests")
-            assert dep.to_requirements_string() == "requests"
+            assert dep.to_requirement_string() == "requests"
 
         def test_single_extra(self):
             dep = Dependency(name="requests", extras=frozenset({"security"}))
-            assert dep.to_requirements_string() == "requests[security]"
+            assert dep.to_requirement_string() == "requests[security]"
 
         def test_multiple_extras(self):
             dep = Dependency(name="requests", extras=frozenset({"security", "socks"}))
-            assert dep.to_requirements_string() == "requests[security,socks]"
+            assert dep.to_requirement_string() == "requests[security,socks]"
